requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
Flask-Mail==0.9.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0
cryptography==41.0.7
qrcode==7.4.2
Pillow==10.0.1
APScheduler==3.10.4
pytz==2023.3
pytest==7.4.3
pytest-flask==1.3.0

# Code quality and formatting tools
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

